original_experiment/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
fake-useragent==0.1.11
future==0.17.1
huepy==0.9.8.1
idna==2.8
instabot==0.20.0
instaloader==4.2.2
itsdangerous==1.1.0
joblib==0.14.1
liac-arff==2.4.0
numpy==1.18.1
pandas==0.25.3
PySocks==1.6.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.21.0
requests-toolbelt==0.9.1
responses==0.10.5
schedule==0.6.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.12.0
tqdm==4.31.1
urllib3==1.24.1
